refactor endpoints /transactions, /fee, /rawtx
