+++ src/module.api/_module.ts
@@ -2,7 +2,7 @@ import { APP_INTERCEPTOR, APP_PIPE } from '@nestjs/core'
 import { CacheModule, Module } from '@nestjs/common'
 import { RpcController } from '@src/module.api/rpc.controller'
 import { ActuatorController } from '@src/module.api/actuator.controller'
-import { TransactionsController } from '@src/module.api/transaction.controller'
+import { TransactionController } from '@src/module.api/transaction.controller'
 import { ApiValidationPipe } from '@src/module.api/pipes/api.validation.pipe'
 import { AddressController } from '@src/module.api/address.controller'
 import { PoolPairController } from '@src/module.api/poolpair.controller'
@@ -11,14 +11,16 @@ import { DeFiDCache } from '@src/module.api/cache/defid.cache'
 import { SemaphoreCache } from '@src/module.api/cache/semaphore.cache'
 import { ExceptionInterceptor } from '@src/module.api/interceptors/exception.interceptor'
 import { ResponseInterceptor } from '@src/module.api/interceptors/response.interceptor'
-import { TokensController } from '@src/module.api/token.controller'
+import { TokenController } from '@src/module.api/token.controller'
 import { BlockController } from '@src/module.api/block.controller'
-import { MasternodesController } from '@src/module.api/masternode.controller'
+import { MasternodeController } from '@src/module.api/masternode.controller'
 import { ConfigService } from '@nestjs/config'
 import { NetworkName } from '@defichain/jellyfish-network'
-import { OraclesController } from '@src/module.api/oracles.controller'
-import { PricesController } from '@src/module.api/prices.controller'
+import { OracleController } from '@src/module.api/oracle.controller'
+import { PriceController } from '@src/module.api/price.controller'
 import { StatsController } from '@src/module.api/stats.controller'
+import { FeeController } from '@src/module.api/fee.controller'
+import { RawtxController } from '@src/module.api/rawtx.controller'
 
 /**
  * Exposed ApiModule for public interfacing
@@ -29,14 +31,16 @@ import { StatsController } from '@src/module.api/stats.controller'
     RpcController,
     AddressController,
     ActuatorController,
-    TransactionsController,
-    TokensController,
+    TransactionController,
+    TokenController,
     PoolPairController,
-    MasternodesController,
+    MasternodeController,
     BlockController,
-    OraclesController,
-    PricesController,
-    StatsController
+    OracleController,
+    PriceController,
+    StatsController,
+    FeeController,
+    RawtxController
   ],
   providers: [
     { provide: APP_PIPE, useClass: ApiValidationPipe },
