+++ src/module.api/block.controller.ts
@@ -1,8 +1,6 @@
 import { Controller, Get, Param, Query } from '@nestjs/common'
 import { Block, BlockMapper } from '@src/module.model/block'
 import { Transaction, TransactionMapper } from '@src/module.model/transaction'
-import { TransactionVinMapper, TransactionVin } from '@src/module.model/transaction.vin'
-import { TransactionVoutMapper, TransactionVout } from '@src/module.model/transaction.vout'
 import { ApiPagedResponse } from '@src/module.api/_core/api.paged.response'
 import { PaginationQuery } from '@src/module.api/_core/api.query'
 
@@ -20,13 +18,11 @@ export function isSHA256Hash (str: string | undefined): boolean {
 export class BlockController {
   constructor (
     protected readonly blockMapper: BlockMapper,
-    protected readonly transactionMapper: TransactionMapper,
-    protected readonly transactionVinMapper: TransactionVinMapper,
-    protected readonly transactionVoutMapper: TransactionVoutMapper
+    protected readonly transactionMapper: TransactionMapper
   ) {
   }
 
-  @Get('')
+  @Get()
   async list (
     @Query() query: PaginationQuery
   ): Promise<ApiPagedResponse<Block>> {
@@ -59,29 +55,5 @@ export class BlockController {
     return ApiPagedResponse.of(transactions, query.size, transaction => {
       return transaction.id
     })
-  }
-
-  @Get('/:hash/transactions/:txid/vins')
-  async getVins (@Param('hash') hash: string, @Param('txid') txid: string, @Query() query: PaginationQuery): Promise<ApiPagedResponse<TransactionVin>> {
-    return ApiPagedResponse.of(await this.getVectors<TransactionVin>(this.transactionVinMapper, hash, txid, query), query.size, vout => {
-      return vout.id
-    })
-  }
-
-  @Get('/:hash/transactions/:txid/vouts')
-  async getVouts (@Param('hash') hash: string, @Param('txid') txid: string, @Query() query: PaginationQuery): Promise<ApiPagedResponse<TransactionVout>> {
-    return ApiPagedResponse.of(await this.getVectors<TransactionVout>(this.transactionVoutMapper, hash, txid, query), query.size, vout => {
-      return vout.n.toString()
-    })
-  }
-
-  async getVectors<T extends TransactionVin | TransactionVout> (mapper: TransactionVinMapper | TransactionVoutMapper, hash: string, txid: string, query: PaginationQuery): Promise<T[]> {
-    const transaction = await this.transactionMapper.get(txid)
-
-    if (transaction !== undefined && transaction.block.hash === hash) {
-      return await mapper.query(txid, query.size) as T[]
-    }
-
-    return []
   }
 }
